+++ src/d-avatar-card2/DAvatarCard2.vue
@@ -2,9 +2,9 @@
   <d-box class="ui-avatar-card2">
     <d-box class="ui-avatar-card2__header">
       <slot name="header">
-        <d-text class="ui-avatar-card2__title" font-face="circularSTD">{{
-          heading
-        }}</d-text>
+        <d-text my0 class="ui-avatar-card2__title" font-face="circularSTD"
+          >{{ heading }}
+        </d-text>
         <slot name="header-button">
           <d-button
             color-scheme="outline"
@@ -18,12 +18,15 @@
     <d-box class="ui-avatar-card2__body">
       <slot name="body">
         <d-box class="ui-avatar-card2__body__left">
-          <d-box class="ui-avatar-card2__title" font-face="circularSTD">{{
-            title
-          }}</d-box>
-          <d-box class="ui-avatar-card2__body__text" font-face="circularSTD"
+          <d-text my0 class="ui-avatar-card2__title" font-face="circularSTD"
+            >{{ title }}
+          </d-text>
+          <d-text
+            my0
+            class="ui-avatar-card2__body__text"
+            font-face="circularSTD"
             >{{ description }}
-          </d-box>
+          </d-text>
         </d-box>
         <d-box class="ui-avatar-card2__body__left">
           <slot name="avatar">
@@ -185,9 +188,5 @@ const emit = defineEmits(["header-button-clicked", "footer-button-clicked"]);
       //border: 1px solid rgba(0, 0, 0, 0.05); //background: linear-gradient( //    0deg, //    rgba(0, 0, 0, 0.15) 0%, //    rgba(0, 0, 0, 0.15) 100% //  ), //  var(--light); //box-shadow: 0px 1px 0px 0px rgba(27, 31, 35, 0.05);
     }
   }
-}
-
-p {
-  margin: 0;
 }
 </style>