+++ changelog.md
@@ -9,7 +9,7 @@
     - Added Console card 2 Component.
 - **09th October, 2023 - Bernode 🙄**
     - Updated components made by me to include a unique class names.
-    - Updated the Continous List Item component and added v-bind to slot.
+    - Updated the Continuous List Item component and added v-bind to slot.
     - Updated Actionable Avatar component button to a link.
 - **26th September, 2023 - Bernode 🙄**
     - Updated the Application Integration component to include button size props and made the icon/logo a loopable slot.